move to manual
