+++ ydb/core/mind/hive/ya.make
@@ -116,4 +116,5 @@ END()
 
 RECURSE_FOR_TESTS(
     ut
+    ut_manual
 )
